Issue 1377 - Break out screenshot tests into individual files
- Break out screenshot tests into individual files
- Add reinstall_app between tests to ensure a clean slate

Issue 1377 - Cleanup with previous commit

Missing newlines

Issue 1377 - More cleanup

Unused import cleanup
